// src/app/generated/models/voice-search-result.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { VoiceSearchCandidate } from '../models/voice-search-candidate';
export interface VoiceSearchResult {
  asrConfidence: number;
  candidates: Array<VoiceSearchCandidate>;
  inconclusive: boolean;
  transcription: string;
}